+++ backend/onyx/db/index_attempt_metrics.py
@@ -1,27 +1,24 @@
-"""Stage-level metric primitives for an `IndexAttempt`.
+"""Stage-level metric write helpers for an `IndexAttempt`.
 
-This module owns the canonical taxonomy of pipeline stages we measure during
-docfetching + docprocessing, the per-stage scope that the API/UI use to
-decide how to display each stage, and the write helpers used by the
-indexing pipeline to record stage timings.
+The canonical pipeline-stage taxonomy (``IndexAttemptStage``,
+``StageScope``, ``STAGE_SCOPE``) lives in
+``onyx.db.index_attempt_metrics_models`` so that ``onyx.db.models`` can
+import the enum as a column type without creating a circular dependency
+through this module. This module is the home of the *runtime* helpers
+that record stage events — the upsert, the timing context manager, and
+the in-memory aggregation buffer.
 
-The ordering of `IndexAttemptStage` members is intentional and represents
-the natural pipeline order — the API surfaces this order to the frontend so
-it can render the "Pipeline order" sort mode without duplicating the
-sequence client-side.
-
-Recording must NEVER fail an indexing attempt. The consumer-facing helpers
-(``time_stage`` and ``StageEventBuffer.flush``) wrap the underlying upsert
-in a try/except that logs and swallows. The lower-level
-``record_stage_aggregate`` and ``record_single_event`` raise on failure so
-they remain testable in isolation.
+Recording must NEVER fail an indexing attempt. The consumer-facing
+helpers (``time_stage`` and ``StageEventBuffer.flush``) wrap the
+underlying upsert in a try/except that logs and swallows. The lower-level
+``record_stage_aggregate`` and ``record_single_event`` raise on failure
+so they remain testable in isolation.
 """
 
 import datetime
 import time
 from collections.abc import Generator
 from contextlib import contextmanager
-from enum import Enum as PyEnum
 
 from sqlalchemy import case
 from sqlalchemy import cast
@@ -31,6 +28,8 @@ from sqlalchemy.dialects.postgresql import insert as pg_insert
 from sqlalchemy.orm import Session
 
 from onyx.db.engine.sql_engine import get_session_with_current_tenant
+from onyx.db.index_attempt_metrics_models import IndexAttemptStage
+from onyx.db.models import IndexAttemptStageMetric
 from onyx.utils.logger import setup_logger
 
 logger = setup_logger()
@@ -40,90 +39,6 @@ logger = setup_logger()
 # meaningful and would corrupt aggregates; protect against monotonic clock
 # weirdness that could in theory yield a negative delta.
 _MIN_DURATION_MS = 0
-
-
-class IndexAttemptStage(str, PyEnum):
-    """Canonical pipeline stages for an `IndexAttempt`.
-
-    Member declaration order matches the natural temporal order of the
-    pipeline (docfetching first, then docprocessing). Both backend
-    serialization and frontend "Pipeline order" display rely on this.
-    """
-
-    # --- Docfetching (one process per attempt) ---
-    CONNECTOR_VALIDATION = "CONNECTOR_VALIDATION"
-    PERMISSION_VALIDATION = "PERMISSION_VALIDATION"
-    CHECKPOINT_LOAD = "CHECKPOINT_LOAD"
-    CONNECTOR_FETCH = "CONNECTOR_FETCH"
-    HIERARCHY_UPSERT = "HIERARCHY_UPSERT"
-    DOC_BATCH_STORE = "DOC_BATCH_STORE"
-    DOC_BATCH_ENQUEUE = "DOC_BATCH_ENQUEUE"
-
-    # --- Docprocessing (one task per batch, many tasks per attempt) ---
-    QUEUE_WAIT = "QUEUE_WAIT"
-    DOCPROCESSING_SETUP = "DOCPROCESSING_SETUP"
-    BATCH_LOAD = "BATCH_LOAD"
-    DOC_DB_PREPARE = "DOC_DB_PREPARE"
-    IMAGE_PROCESSING = "IMAGE_PROCESSING"
-    CHUNKING = "CHUNKING"
-    CONTEXTUAL_RAG = "CONTEXTUAL_RAG"
-    EMBEDDING = "EMBEDDING"
-    VECTOR_DB_WRITE = "VECTOR_DB_WRITE"
-    POST_INDEX_DB_UPDATE = "POST_INDEX_DB_UPDATE"
-    COORDINATION_UPDATE = "COORDINATION_UPDATE"
-
-    # --- Aggregate ---
-    BATCH_TOTAL = "BATCH_TOTAL"
-
-
-class StageScope(str, PyEnum):
-    """How often a stage fires per attempt.
-
-    Used by the admin UI to decide where to render each stage:
-
-    - `ATTEMPT_LEVEL` stages have at most one event per attempt and are
-      shown as a small "Attempt overhead" disclosure.
-    - `BATCH_LEVEL` stages have many events per attempt and contribute to
-      the per-batch stacked bar / detail table.
-    """
-
-    ATTEMPT_LEVEL = "ATTEMPT_LEVEL"
-    BATCH_LEVEL = "BATCH_LEVEL"
-
-
-# Single source of truth for stage scope. The API serializes this onto each
-# row so the frontend never has to duplicate the mapping.
-STAGE_SCOPE: dict[IndexAttemptStage, StageScope] = {
-    IndexAttemptStage.CONNECTOR_VALIDATION: StageScope.ATTEMPT_LEVEL,
-    IndexAttemptStage.PERMISSION_VALIDATION: StageScope.ATTEMPT_LEVEL,
-    IndexAttemptStage.CHECKPOINT_LOAD: StageScope.ATTEMPT_LEVEL,
-    IndexAttemptStage.CONNECTOR_FETCH: StageScope.BATCH_LEVEL,
-    IndexAttemptStage.HIERARCHY_UPSERT: StageScope.BATCH_LEVEL,
-    IndexAttemptStage.DOC_BATCH_STORE: StageScope.BATCH_LEVEL,
-    IndexAttemptStage.DOC_BATCH_ENQUEUE: StageScope.BATCH_LEVEL,
-    IndexAttemptStage.QUEUE_WAIT: StageScope.BATCH_LEVEL,
-    IndexAttemptStage.DOCPROCESSING_SETUP: StageScope.BATCH_LEVEL,
-    IndexAttemptStage.BATCH_LOAD: StageScope.BATCH_LEVEL,
-    IndexAttemptStage.DOC_DB_PREPARE: StageScope.BATCH_LEVEL,
-    IndexAttemptStage.IMAGE_PROCESSING: StageScope.BATCH_LEVEL,
-    IndexAttemptStage.CHUNKING: StageScope.BATCH_LEVEL,
-    IndexAttemptStage.CONTEXTUAL_RAG: StageScope.BATCH_LEVEL,
-    IndexAttemptStage.EMBEDDING: StageScope.BATCH_LEVEL,
-    IndexAttemptStage.VECTOR_DB_WRITE: StageScope.BATCH_LEVEL,
-    IndexAttemptStage.POST_INDEX_DB_UPDATE: StageScope.BATCH_LEVEL,
-    IndexAttemptStage.COORDINATION_UPDATE: StageScope.BATCH_LEVEL,
-    IndexAttemptStage.BATCH_TOTAL: StageScope.BATCH_LEVEL,
-}
-
-
-# Defensive sanity check: every enum member must have a scope. This runs at
-# import time so a future contributor adding a stage but forgetting the
-# scope mapping fails fast rather than at API serialization.
-_missing_scopes = set(IndexAttemptStage) - set(STAGE_SCOPE)
-if _missing_scopes:
-    raise RuntimeError(
-        f"IndexAttemptStage members missing from STAGE_SCOPE: {_missing_scopes}"
-    )
 
 
 def record_stage_aggregate(
@@ -153,10 +68,6 @@ def record_stage_aggregate(
     """
     if event_count <= 0:
         return
-
-    # Imported here to avoid a circular dependency: ``onyx.db.models`` imports
-    # ``IndexAttemptStage`` from this module.
-    from onyx.db.models import IndexAttemptStageMetric
 
     if now is None:
         now = datetime.datetime.now(datetime.timezone.utc)
