+++ backend/tests/external_dependency_unit/db/test_index_attempt_stage_metrics.py
@@ -30,10 +30,10 @@ from sqlalchemy.orm import Session
 
 from onyx.db.engine.sql_engine import get_session_with_current_tenant
 from onyx.db.enums import IndexingStatus
-from onyx.db.index_attempt_metrics import IndexAttemptStage
 from onyx.db.index_attempt_metrics import record_single_event
 from onyx.db.index_attempt_metrics import record_stage_aggregate
 from onyx.db.index_attempt_metrics import StageEventBuffer
+from onyx.db.index_attempt_metrics_models import IndexAttemptStage
 from onyx.db.models import ConnectorCredentialPair
 from onyx.db.models import IndexAttempt
 from onyx.db.models import IndexAttemptStageMetric
